Ep3Logica v1.0.0 Ep3Logica

Implementação de um automato finito aceitador de estados deterministico e não deterministico.

Types

action()

action() :: atom()

automaton()

automaton() :: %{
  current_state: state(),
  transitions: transitions(),
  accept_states: [state(), ...],
  tape: tape()
}

state()

state() :: atom()

tape()

tape() :: %{left: [action(), ...], right: [action(), ...]}

transitions()

transitions() :: [{state(), [{action(), [state()]}, ...]}, ...]

Functions

create_automaton(initial_state, transitions, accept_states, actions)

create_automaton(state(), transitions(), [state(), ...], [action(), ...]) ::
  automaton()

Cria o automato finito

do_step(automaton, new_state)

do_step(automaton(), state()) :: automaton()

Muda o estado do automato para "new_state" e reconfigura sua fita.

get_next_states(automaton)

get_next_states(automaton()) :: [state(), ...]

Retorna uma lista com os possíveis próximos estados do automato.

is_accepted?(automaton)

is_accepted?(automaton()) :: true | false

Verifica se o automato se encontra em um estado de aceitação

run_automaton(automaton)

run_automaton(automaton()) :: true | false

Executa o automato finito deterministico ou não deterministico. Tratando-se de um automato "aceitador" busca-se um estado final que seja um estado de aceitação. Caso qualquer um dos caminhos gerados pela sequência de ações resulte em um estado de aceitação a função retorna true, caso contrário retorna false.