Ep3Logica v1.0.0 Tape

Unidade de comunicação de um automato finito baseado em Teaching Nondeterministic and Universal Automata using Scheme (Christian Wagenknecht and Daniel P Friedman)

Types

action()

action() :: atom()

tape()

tape() :: %{left: [action(), ...], right: [action(), ...]}

Functions

at(map)

at(tape()) :: action()

Retorna o primeiro elemento da lista a direita, onde está localizada a "cabeça" de leitura da fita

contents(tape)

contents(tape()) :: tape()

Apresenta o conteúdo da fita

end?(tape)

end?(tape()) :: true | false

Verifica se é o final da fita comparando com o átomo :"$"

init(actions)

init([action(), ...]) :: tape()

Inicializa a fita com uma lista vazia a esquerda e a lista de valores recebidos a direita

reconfig(map)

reconfig(tape()) :: tape()

Move a cabeça de leitura uma celula a direita, adicionando o valor movido a lista da esquerda